ToolCard refactor base structure
