+++ client/src/style/scss/overrides.scss
@@ -35,7 +35,7 @@ pre {
 .dropdown-menu {
     .dropdown-item {
         text-decoration: none;
-        .fa {
+        .fa, .svg-inline--fa {
             @extend .fa-fw;
             @extend .mr-2;
         }
